Initial commit
Create a json socket class to handle sending and recieving json data.
The message format is the length of the json encoded string followed by
the json encoded string.

Initial use is to run init the server with server_driver.py then send
a test message with client_driver.py. The server echoes back the
message.